remove some time sensitive tests from under -race testing
