+++ etcd/etcd_test.go
@@ -1,30 +1,14 @@
-// Copyright 2015 Zalando SE
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package etcd
 
 import (
 	"encoding/base64"
 	"errors"
 	"log"
-	"net"
 	"net/http"
 	"net/http/httptest"
 	"os"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/zalando/skipper/eskip"
 	"github.com/zalando/skipper/etcd/etcdtest"
@@ -195,44 +179,6 @@ func TestFailedEndpointsRotation(t *testing.T) {
 
 	if err == nil {
 		t.Error("failed to fail")
-	}
-
-	expectedEndpoints := []string{s2.URL, "neverreached", s.URL}
-
-	if strings.Join(c.endpoints, ";") != strings.Join(expectedEndpoints, ";") {
-		t.Error("wrong endpoints rotation")
-	}
-}
-
-func TestTimedoutEndpointsRotation(t *testing.T) {
-	s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-	}))
-
-	s.Close()
-
-	s2 := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		time.Sleep(time.Duration(5 * time.Millisecond))
-	}))
-	defer s2.Close()
-
-	c, err := New(Options{Endpoints: []string{s.URL, s2.URL, "neverreached"}, Prefix: "/skippertest-invalid"})
-	c.client.Timeout = time.Duration(1 * time.Millisecond)
-
-	if err != nil {
-		t.Error(err)
-		return
-	}
-
-	_, err = c.LoadAll()
-
-	if err == nil {
-		t.Error("failed to fail")
-	}
-
-	nerr, ok := err.(net.Error)
-
-	if !ok || !nerr.Timeout() {
-		t.Error("timeout error expected")
 	}
 
 	expectedEndpoints := []string{s2.URL, "neverreached", s.URL}
